api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
prometheus-client==0.21.0
pydantic==2.10.1
pydantic[email]==2.10.1
motor==3.6.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
httpx==0.27.2
# Dev/test (excluded from prod image via requirements-test.txt)
